+++ rocketcea/cea_obj.py
@@ -335,7 +335,8 @@ class CEA_Obj(object):
 
 
     def setupCards(self, Pc=100.0, MR=1.0, eps=40.0, PcOvPe=None, frozen=0,
-                   ERphi=None, ERr=None, frozenAtThroat=0, short_output=0):
+                   ERphi=None, ERr=None, frozenAtThroat=0, short_output=0,
+                   show_transport=0):
         '''
         Set up card deck and call CEA FORTRAN code.::
 
@@ -407,8 +408,12 @@ class CEA_Obj(object):
             N += 1
 
         for line in ["   ","outp   calories ","   ","end "]:
-            if (line=="outp   calories ") and short_output:
-                line = "outp   calories short "
+            if (line=="outp   calories "):
+                if short_output:
+                    line += " short "
+                if show_transport:
+                    line += " transport "
+                    
             set_py_cea_line(N,line)
             N += 1
 
@@ -469,14 +474,17 @@ class CEA_Obj(object):
         #print( '"'+self.pathPrefix+'"',' and ', '"'+myfile+'"' )
         py_cea.py_cea(self.pathPrefix, myfile, self.makeOutput, readData)
 
-    def get_full_cea_output(self, Pc=100.0, MR=1.0, eps=40.0, frozen=0, frozenAtThroat=0, short_output=0):
+    def get_full_cea_output(self, Pc=100.0, MR=1.0, eps=40.0, frozen=0, 
+                            frozenAtThroat=0, short_output=0, show_transport=1):
         """Get the full output file created by CEA. Return as a string."""
 
         # regardless of how run was set up, change makeOutput flag True
         save_flag = self.makeOutput
         self.makeOutput = True
 
-        self.setupCards( Pc=Pc, MR=MR, eps=eps, frozen=frozen, frozenAtThroat=frozenAtThroat, short_output=short_output)
+        self.setupCards( Pc=Pc, MR=MR, eps=eps, frozen=frozen, 
+                         frozenAtThroat=frozenAtThroat, short_output=short_output,
+                         show_transport=show_transport)
 
         self.makeOutput = save_flag # restore makeOutput
 
@@ -728,6 +736,20 @@ class CEA_Obj(object):
         self.setupCards( Pc=Pc, MR=MR, eps=eps)
         M = py_cea.rockt.vmoc[2]
         return M
+
+    def get_Temperatures(self, Pc=100.0, MR=1.0,eps=40.0):
+        """::
+
+        #: Return a list of temperatures at the chamber, throat and exit.
+        #: (Note Texit is equilibrium temperature NOT Frozen temperature)
+        #: MR is only used for ox/fuel combos.
+        """
+        self.setupCards( Pc=Pc, MR=MR, eps=eps)
+        
+        # convert from Kelvin to Rankine
+        tempList = 1.8 * py_cea.prtout.ttt[:3]
+        return tempList # Tc, Tthroat, Texit
+
 
     def get_SonicVelocities(self, Pc=100.0, MR=1.0,eps=40.0):
         """::